+++ CHANGELOG.md
@@ -10,8 +10,8 @@ It is planned to support 7.8.1 when this version is released.
 ### v1.0.2 (2022-09-21)
 
 - Do not run extra checks in `host-check` by default.
-- Make igb_uio a supported PCI driver, and only fail `host-check` if no interface driver is loaded.
-- Only check if IOMMU is enabled if vfio-pci is being used, and handle the case where IOMMU is unconfigurable.
+- Make igb_uio a supported PCI driver, only failing `host-check` if no interface driver is loaded.
+- Only check if IOMMU is enabled if vfio-pci is being used, and handle the case where the vfio-pci 'no IOMMU' mode is unconfigurable.
 
 
 ### v1.0.1 (2022-09-07)